requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
einops>=0.7.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
rasterio>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
earthengine-api>=0.1.380
geemap>=0.30.0
wandb>=0.16.0
statsmodels>=0.14.0
